Refactor classes `Struct`, `Service`, and `Job` to `mqttwarn.model`
