+++ mqttwarn/testing/fixtures.py
@@ -2,7 +2,7 @@ import logging
 
 import pytest
 
-from mqttwarn.core import Service
+from mqttwarn.model import Service
 
 
 @pytest.fixture
@@ -11,4 +11,5 @@ def mqttwarn_service():
     A service instance for propagating to the plugin.
     """
     logger = logging.getLogger(__name__)
-    return Service(mqttc=None, logger=logger)
+    # FIXME: Should propagate `mqttwarn.core.globals()` to `mwcore`.
+    return Service(mqttc=None, logger=logger, mwcore={}, program="mqttwarn-testdrive")
